'size_t' is not the same as 'unsigned long'
So using %lu to printf 'size_t' is not 100% portable either. At least in
64-bit Windows code 'size_t' is 64 bits but 'unsigned long' is 32 bits.

Let's just use 'unsigned' for the loop indexes here (and %u as the
format). After all, they are single-digit numbers.

Alternatively, could just use C++ stream output for this purely informative
printout.

+++ sc/qa/unit/ucalc.cxx
@@ -6243,16 +6243,16 @@ void Test::testFormulaGrouping()
 
     m_pDoc->InsertTab( 0, "sheet" );
 
-    for (size_t i = 0; i < SAL_N_ELEMENTS( aGroupTests ); i++)
+    for (unsigned i = 0; i < SAL_N_ELEMENTS( aGroupTests ); i++)
     {
-        for (size_t j = 0; j < SAL_N_ELEMENTS( aGroupTests[0].pFormula ); j++)
+        for (unsigned j = 0; j < SAL_N_ELEMENTS( aGroupTests[0].pFormula ); j++)
         {
             OUString aFormula = OUString::createFromAscii(aGroupTests[i].pFormula[j]);
             m_pDoc->SetString(0, (SCROW)j, 0, aFormula);
         }
         m_pDoc->RebuildFormulaGroups();
 
-        for (size_t j = 0; j < SAL_N_ELEMENTS( aGroupTests[0].pFormula ); j++)
+        for (unsigned j = 0; j < SAL_N_ELEMENTS( aGroupTests[0].pFormula ); j++)
         {
             ScBaseCell *pCell = NULL;
             m_pDoc->GetCell( 0, (SCROW)j, 0, pCell );
@@ -6268,7 +6268,7 @@ void Test::testFormulaGrouping()
 
             if( !!pCur->GetCellGroup().get() ^ aGroupTests[i].bGroup[j] )
             {
-                printf("expected group test %lu at row %lu to be %d but is %d\n",
+                printf("expected group test %u at row %u to be %d but is %d\n",
                        i, j, aGroupTests[i].bGroup[j], !!pCur->GetCellGroup().get());
                 CPPUNIT_ASSERT_MESSAGE("Failed", false);
             }
